refactor: componentize the home
