+++ src/views/pages/home/components/statistics-modal.tsx
